Add workflow to check mirrors are functional
This Actions workflow will run twice every day. It checks ziglang.org
for the latest `master` build, downloads it from ziglang.org and runs it
through sha512sum, and does the same for each mirror. If the download
from the mirror fails, or the sha512sum output differs, it sends me
personally a notification via an external webhook.